Upload source code to github from quiz4
The program is about calculating PI in a multithreaded environment.

Makefile: Using gcc compiler with options:

  - Option -Wall for enable all warnings

  - Option -pthread for #include <pthread.h> in thread.[ch]

  - Option -lm for #include <math.h> in main.c

leibniz.h: The implementation for Leibniz’s Formula for PI.

thread.[ch]: The implementation of a thread pool.

main.c: Using Bailey–Borwein–Plouffe formula and the thread pool
to evaluate the value of PI.